compiler: Introduce an executable_line2 instruction
In order to support breakpoints on lines, we need to know which
X registers are live at that point. For this, we replace the
`executable_line` instruction by a new version that contains
also the information on live registers.

This information is encoded in a 32-bit integer as follows:

  - If the MSB is set, then the remaining 31 bits are interpreted
    as a bitset, with bit 0 meaning "X0 is alive", etc.
  - If the MSB is not set, then it must be a value 0 <= N <= 1024,
    and denote that all registers X0...X{N-1} are live.

At this point the compiler doesn't use a bitset internally nor
guarantees that if more than 31 vars are live at the same time,
there are no gaps, but this servers as an interface we can use
on the VM.

We could in the future use a 64-bit integer instead, but the
module loader doesn't handle 64-bit constants as Beam op args
right now