erl_debugger: Add skeleton module
We add a new erl_debugger module to the kernel
application, that will provides an API for writing
debuggers. Here we set up the general interface,
with a stub for "setting a breakpoint" as the only
debugging primitive. In the next commits we first
implement the "breakpoint" primitive, and later
add additional ones.

About the general design:

- A node needs to have debugging support enabled
  on startup with the `+D` option and this can't be
  enabled afterwards. This gives admins a way to
  prevent processes from blocking on prod due to an
  accidental debugger usage, etc.
- Debugging primitives based on code instrumentation
  can be individually enabled/disabled, etc.
- A debugger is an Erlang process; there can be at most
  one debugger process per node.
- The system will communicate with the debugger process
  via `debugger_events` messages. For now we introduce
  only one message for notifying that a process hit
  a breakpoint