[pause_proc_timer][2/n] Allow system-tasks to be scheduled while holding proc_status lock
In the next commit, we will want to schedule a system-task to resume a paused
timer from within `resume_process()`. At this point, we will be holding the
proc_status lock, and scheduling a system-task, currently tries to acquire this
lock unconditionally.

So now, when scheduling a system-task we pass the locks that are currently held,
so that `active_sys_enqueue()` can decide whether to take the lock or not.