beamasm: Factor out logic to flip breakpoint trampolines
We want to reuse the logic currently use to flip beakpoints for tracing, etc.
However, at the moment, the functions where this happens not only change the
code, but also set some flags value that are specific to those cases..

We factor out this part of the code to their own functions. In doing so, we observe
that for x86-64, the code was referring to `BEAM_ASM_FUNC_PROLOGUE_SIZE` to determine
the offset to use when "disabling" a breakpoint that doesn't seem relevant. The value
is indeed 6 (see, e.g., the assertion on the "enabling" case), and this corresponds to:
  - 1 byte for the NOP 0x90 due to call alignment
  - 1 byte for the opcode of a `CALL`
  - 4 bytes for the 32-bit offset of the call target