[pause_proc_timer][4/n] Use a 64-bit state in ErtsMonitorSuspend
In the next commit, we will add a flag to the process-suspend monitor,
to indicate whether the timer was paused as well. Essentially, we want
the second-most-significant bit to denote the new flag and the rest of
the bits to be the counter.

The problem is that on 32-bits architectures the size of the counter may
not be large enough. The docs for `erlang:suspend_process/2` say:

> Failures:
> ...
>
> **system_limit**
> If the process identified by Suspendee has been suspended more times
> by the calling process than can be represented by the currently used
> internal data structures. The system limit is greater than 2,000,000,000
> suspends and will never be lower.

That is, we need to guarantee that we have at least 31 bits for the
counter in the monitor state. However, on 32-bit architectures, the
monitor state will be a 32-bit wide, and we don't have enough space for
two flags.

So here we make the process-suspend monitor state 64-bit wide on all
architectures. On modern machines this change is a no-op. On 32-bit
architecures we are adding one word to the structure, which given alignment,
is the minimum possible increase.