[pause_proc_timer][6/n] Add pause_proc_timer option to suspend_process/2
We add a new `pause_proc_timer` option to the `erlang:suspend_process/2`
BIF. When given, the process is not only suspended, but its proc timer,
if set, will be paused, and resumed when the process is resumed.

This means that if the paused process is waiting on a `receive`, it will not
timeout even if suspended for long.

We add testcases for this functionality