Support coverage also when using beam_debug_info
On module loading, we convert `debug_line` instructions into
`executable_line` instructions. However, if debugger support
is enabled (and the line-breakpoint) instrumentation, we also
emit a `i_debug_line` instruction, that will handle the
instrumentation (to be introduced next).

In this case, we will attempt to add a line-entry twice, but
the second one (coming from the executable_line instruction)
will be ignored as duplicate. However, we now delay the
the overflow until the point where we actually try to add an
entry, or we'll get spurious failures to load (e.g. when
no_line_info is used)