compiler: Use a map for tracking debug_info
At the moment, we only expose frame-size and variable-mappings
when compiling with `+beam_debug_info`. Internally, they are
stored as a tuple.

We are planning to add more debug-info, but this will mean changing
all places where the tuple is used, to turn it into a triple, and
the same will happen if we later add another one, etc.

So to make this refactorings easier, we introduce a map.