Make code:get_debug_info/1 use a map for the debug info
At the moment, we are only saving frame-size and var mappings
as debug-info, but we are planning to add more information.
Because code:get_debug_info/1 returns the current info in a
tuple, it will not be possible to add more fields without
making breaking changes.

So instead use a map `#{frame_size => ..., vars => ...}`, so that
new fields can be added in the future by extending this map