Fix suspend_process() issue on processes in dirty schedulers
# Context
`erlang:suspend_process()` was broken when applied on processes that
happened to be in a dirty-scheduler. The symptom was:
  - `internal_error` would be returned when called synchronously
  - `not_suspended` would be delivered when called asynchronously,
     even though the process was in fact suspended

# Problem
When trying to suspend a processes in a dirty-scheduler, it is actually
the dirty-scheduler who suspends the process, by calling
`erts_proc_sig_handle_pending_suspend()`. This essentially activates all
pending suspend monitors and finally sends all pending sync messages, by
calling `sync_suspend_reply()`.

The problem is that when calling the latter, it is passing the process
state *before* activating the monitors. In particular, the SUSPENDED
flag is not set, what makes `sync_suspend_reply()` assume that something
is wrong.