[buck2] Add rule to build resources from priv/ dirs
Some of the apps need resources. The problem is that they
are stored in a priv/ directory, and there is currently no
easy way to give them to the `resources` attr of the
`erlang_application()` rule, and not have an extra `priv/`
directory under the app's `priv/` directory

So we add a simple rule to build resources that will drop
any pre-existing `priv/` dir. It is a bit of a hack, but well