Fix race condition in bif timer suspension
When suspending a process we pause all bif timers targeting it.
But until this diff we did not pause any bif timer created after the
suspension. Which is an issue for the debugger, as suspending all the
processes is not atomic, so new timers can keep appearing even as we're
in the middle of suspending processes.

This makes the test adding in the last commit pass (it was consistently
failing before).