[buck2] Ensure test node is distributed
When running tests with `make`, the `-sname`
arg is passed with `$CT_NODENAME` which defaults
to `test_node`. That the node is alive is then
used by things like ?CT_PEER().

So we add a CT hook that starts net_kernel if needed.
We use `test_server` as name, which is enough since
buck2 uses a different epmd instance per test